deployment mods
decomment any of the comments. if you do it is likely that you wil break the website and not be able to. if you do.
comment the comments. then clear your browser chace then ou will be able to accsse the webiste.
how i know?
i decommented the comments and fixed it (sortof)